Update swipe refresh extension
