+++ app/src/main/java/com/jay/movies/ext/SwipeRefreshExt.kt
@@ -1,0 +1,13 @@
+package com.jay.movies.ext
+
+import androidx.databinding.BindingAdapter
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+
+@BindingAdapter("isRefreshing")
+fun SwipeRefreshLayout.bindRefreshing(isRefreshing: Boolean) {
+    this.isRefreshing = isRefreshing
+}
+
+@BindingAdapter("onRefresh")
+fun SwipeRefreshLayout.bindRefreshListener(onRefreshListener: SwipeRefreshLayout.OnRefreshListener) =
+    setOnRefreshListener(onRefreshListener)
